[libc] Small type fix

+++ user/libc/unistd.cpp
@@ -48,7 +48,7 @@ extern "C" {
 
     ssize_t write(int fd, const void* buffer, size_t count) {
         /// TODO: check return value and set errno.
-        int rc = 0;
+        ssize_t rc = 0;
         while ((rc = syscall<ssize_t>(SYS_write, fd, buffer, count)) == -2);
         return rc;
     }
